refactor(packages) remove dupliated types
